moved methods for filesRow and related element to seperate PO
